requirements.txt
pygame==2.1.2
PyOpenGL==3.1.6
pillow==9.1.0
PyGLM==2.5.7